tests/wpt/web-platform-tests/2dcontext/pixel-manipulation/2d.imageData.object.set.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.imageData.object.set</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="stylesheet" href="/common/canvas-tests.css">
<body class="show_output">

<h1>2d.imageData.object.set</h1>
<p class="desc">ImageData.data can be modified</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("ImageData.data can be modified");
_addTest(function(canvas, ctx) {

var imgdata = ctx.getImageData(0, 0, 10, 10);
imgdata.data[0] = 100;
_assertSame(imgdata.data[0], 100, "imgdata.data[\""+(0)+"\"]", "100");
imgdata.data[0] = 200;
_assertSame(imgdata.data[0], 200, "imgdata.data[\""+(0)+"\"]", "200");


});
</script>
